Move the meat of the module to index.js.
This should make Ryan smile.
